requirements.txt
numpy==1.16.4
keras==2.4.3
matplotlib==3.3.1
Pillow==7.2.0
tensorflow==2.3.0
tkinter==1.0.0
pandas
glob
